+++ include/Map.hpp
@@ -14,14 +14,14 @@ public:
     Map(const Type& map);
     ~Map() = default;
 
-    short int altitude(double latitude, double longitude);
-    double interpolated_altitude(double latitude, double longitude);
+    short int altitude(float latitude, float longitude);
+    float interpolated_altitude(float latitude, float longitude);
 
-    static Type initialize(std::string dem_directory_path, unsigned int nrows, unsigned int ncols, double cellsize, short int nodata);
+    static Type initialize(std::string dem_directory_path, unsigned int nrows, unsigned int ncols, float cellsize, short int nodata);
 
 private:
     DEM dem;
     Type map;
 
-    bool load(double latitude, double longitude);
+    bool load(float latitude, float longitude);
 };
